+++ lib/coglax_screenshot.dart
@@ -1,8 +1,0 @@
-export 'package:coglax_screenshot/src/check_screenshot_runable.dart';
-export 'package:coglax_screenshot/src/screenshot_result.dart';
-export 'package:coglax_screenshot/src/screenshot_result_page.dart';
-export 'package:coglax_screenshot/src/screenshot_service.dart';
-export 'package:coglax_screenshot/src/take_screenshots.dart';
-export 'package:coglax_screenshot/src/types/screenshot_config.dart';
-export 'package:coglax_screenshot/src/types/screenshot_mode.dart';
-export 'package:coglax_screenshot/src/types/screenshot_page_info.dart';