+++ lib/colaxy_screenshot.dart
@@ -1,0 +1,8 @@
+export 'package:colaxy_screenshot/src/check_screenshot_runable.dart';
+export 'package:colaxy_screenshot/src/screenshot_result.dart';
+export 'package:colaxy_screenshot/src/screenshot_result_page.dart';
+export 'package:colaxy_screenshot/src/screenshot_service.dart';
+export 'package:colaxy_screenshot/src/take_screenshots.dart';
+export 'package:colaxy_screenshot/src/types/screenshot_config.dart';
+export 'package:colaxy_screenshot/src/types/screenshot_mode.dart';
+export 'package:colaxy_screenshot/src/types/screenshot_page_info.dart';