+++ lib/src/screenshot_service.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'dart:ui' as ui;
 
-import 'package:coglax_screenshot/coglax_screenshot.dart';
+import 'package:colaxy_screenshot/colaxy_screenshot.dart';
 import 'package:device_frame_plus/device_frame_plus.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart' hide Image;
@@ -28,7 +28,8 @@ class ScreenshotService {
 
   /// スクリーンショットを実行する
   Future<void> executeScreenshots() async {
-    debugPrint('executeScreenshots');
+    final defaultDelay = config.captureDelay;
+    var isFirst = true;
     // 各デバイス × 各言語 × 各ページの組み合わせでスクリーンショットを作成
     for (final mode in ScreenshotModeInfo.all) {
       mode.setWindowToSize();
@@ -37,11 +38,16 @@ class ScreenshotService {
           if (!config.indexToScreenshot.contains(page.index)) {
             continue;
           }
+          if (isFirst) {
+            config.captureDelay = const Duration(seconds: 3);
+            isFirst = false;
+          }
           await _capturePageScreenshot(
             locale: locale,
             page: page,
             modeInfo: mode,
           );
+          config.captureDelay = defaultDelay;
         }
       }
     }