+++ lib/src/types/screenshot_config.dart
@@ -1,10 +1,10 @@
-import 'package:coglax_screenshot/src/types/screenshot_page_info.dart';
+import 'package:colaxy_screenshot/src/types/screenshot_page_info.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 /// スクリーンショット用の設定クラス
 class ScreenshotConfig {
-  const ScreenshotConfig({
+  ScreenshotConfig({
     required this.imghippoApiKey,
     required this.supportedLocales,
     required this.pages,
@@ -33,7 +33,7 @@ class ScreenshotConfig {
   final List<ScreenshotPageInfo> pages;
 
   /// スクリーンショット間の待機時間（デフォルト3秒）
-  final Duration captureDelay;
+  Duration captureDelay;
 
   /// 背景色（デフォルトはダークグレー）
   final Color backgroundColor;