+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+- Fix First screenshot loading issue.
+
+## 0.0.2
+
+- Fix First screenshot loading issue.
+
 ## 0.0.1
 
 - Initial release.